Start anew in french
OK. Starting to write in english was a mistake. It takes me a lot of
time to write in english. So I decided to start again, in french, so
that most of the informations is quickly written and available.

I'll work on the english translation as soon as I'll have something
almost complete in french.

I didn't delete the english pages. I've put them in the i18n/en
directory.